+++ packages/services/ResearchGroup/CHANGELOG.md
@@ -3,9 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.6](https://gitlab.com/DEIP/deip-client-modules/compare/@deip/research-group-service@1.0.5...@deip/research-group-service@1.0.6) (2020-03-18)
+# 1.0.0 (2020-03-18)
 
-**Note:** Version bump only for package @deip/research-group-service
+
+### Bug Fixes
+
+* deipRpc config ([879ce0c](https://gitlab.com/DEIP/deip-client-modules/commit/879ce0c369bdfa741f7044f4b14fa2990c767f99))
+
+
+
+
+
+# 1.0.0 (2020-03-18)
+
+
+### Bug Fixes
+
+* deipRpc config ([879ce0c](https://gitlab.com/DEIP/deip-client-modules/commit/879ce0c369bdfa741f7044f4b14fa2990c767f99))